feat(web): ship minimal waitlist landing as production default (#11296)
Default the homepage to a single 100vh email-capture hero while the iOS
app is pre-launch. Keep the full System B marketing page behind
NEXT_PUBLIC_LYB_FULL_LANDING=1 and persist waitlist signups to
waitlist_entries via /api/waitlist with idempotent upserts.
