+++ apps/web/.env.example
@@ -18,6 +18,10 @@ NEXT_PUBLIC_REVENUECAT_PUBLIC_KEY=your-revenuecat-key
 # Optional: Version (auto-populated in CI)
 NEXT_PUBLIC_VERSION=local-dev
 
+# Landing page mode. Unset/false = minimal 100vh waitlist hero (production default).
+# Set to 1 locally or in preview to render the full System B marketing landing.
+NEXT_PUBLIC_LYB_FULL_LANDING=
+
 # OpenAI API Configuration (for avatar generation)
 # NEVER commit your actual API key to version control
 OPENAI_API_KEY=your-openai-api-key-here
