+++ apps/web/src/app/__tests__/home-page.test.ts
@@ -1,4 +1,5 @@
 import { getPricingPlanNote, landingSectionHeadings } from '../home-page-copy';
+import { waitlistLandingCopy } from '../waitlist-copy';
 
 describe('HomePage landing copy', () => {
   it('keeps the System B section headings wired to the homepage module', () => {
@@ -14,5 +15,10 @@ describe('HomePage landing copy', () => {
   it('returns pricing notes for annual and monthly states', () => {
     expect(getPricingPlanNote(true)).toBe('$5.83/month equivalent. Save 42%.');
     expect(getPricingPlanNote(false)).toBe('Switch anytime when you are ready.');
+  });
+
+  it('keeps waitlist hero copy separate from full-landing section headings', () => {
+    expect(waitlistLandingCopy.headline).not.toEqual(landingSectionHeadings.hero);
+    expect(waitlistLandingCopy.submitLabel).toBe('Join the waitlist');
   });
 });
