+++ apps/web/src/app/waitlist-copy.ts
@@ -1,0 +1,12 @@
+export const waitlistLandingCopy = {
+  headline: 'Track your body composition privately.',
+  subheading:
+    'LogYourBody is coming soon. Join the waitlist for early access when the app launches.',
+  emailLabel: 'Email',
+  emailPlaceholder: 'you@example.com',
+  submitLabel: 'Join the waitlist',
+  successMessage: "You're on the list. We'll email you when LogYourBody launches.",
+  duplicateMessage: "You're already on the waitlist. We'll be in touch.",
+  errorMessage: 'Something went wrong. Please try again.',
+  invalidEmailMessage: 'Enter a valid email address.',
+} as const;
