+++ apps/web/src/lib/waitlist/normalize-email.ts
@@ -1,0 +1,9 @@
+const EMAIL_PATTERN = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+
+export function normalizeWaitlistEmail(raw: string): string | null {
+  const trimmed = raw.trim().toLowerCase();
+  if (!trimmed || !EMAIL_PATTERN.test(trimmed)) {
+    return null;
+  }
+  return trimmed;
+}
